Add an advisory note to the Extra Action Button for users to use Edit Mode instead
Taking a frame out of Edit Mode and letting Bartender4 handle it instead
is problematic at best, and for the Extra Action Button in particular we
don't even do anything to it other then making it movable.

Favoring Edit Mode and avoiding issues with the button not working right
seems like the best path forward, even though I was not able to
reproduce issues in raids or dungeons myself.