Fix the disappearing micro menu
The micro menu is not properly restored in WoW Classic when a vehicle is left. This results in the micro menu disappearing when mouniting the vehicle but not reapearing after leaving it. The root cause seems to be that the if-conditions changed in this commit were never satisfied in WoW classic, so that the micro menu was not restored. Fixes #154 in Wotlk 3.4.1.